implement cronjob to periodically process failed events

+++ metadata/internal/cronjob/cronjob.go
@@ -23,20 +23,22 @@ func NewCronjob(metadataEventRepo driven.IMetadataEventRepository, producer driv
 func (c *Cronjob) Run() {
 	for {
 		events, err := c.metadataEventRepo.Load(context.Background())
-		if err != nil {
+		if err == nil {
+			for _, event := range *events {
+				err = c.producer.PublishMetadataSubmittedEvent(event)
+				if err != nil {
+					log.Println(err)
+					continue
+				}
+				err = c.metadataEventRepo.Delete(context.Background(), event.ID)
+				if err != nil {
+					log.Printf("Saved event with ID %s and record type %s published but could not be cleaned up: %v\n", event.ID, event.RecordType, err)
+				}
+			}
+		} else {
 			log.Println(err)
 		}
 
-		for _, event := range *events {
-			err = c.producer.PublishMetadataSubmittedEvent(event)
-			if err != nil {
-				continue
-			}
-			err = c.metadataEventRepo.Delete(context.Background(), event.ID)
-			if err != nil {
-				log.Printf("Saved event with ID %s and record type %s published but could not be cleaned up: %v\n", event.ID, event.RecordType, err)
-			}
-		}
 		time.Sleep(1 * time.Minute)
 	}
 }
